User functionaliy and employee functionality fully complete, working on admin

+++ src/main/java/com/revature/Transferable.java
@@ -1,5 +1,5 @@
 package com.revature;
 
 public interface Transferable {
-	public void Transfer(int transferee, int amount);
+	public void Transfer(int transferee, double amount);
 }
